Move from gometalinter to golangci-lint
gometalinter has been deprecated so we move to golangci-lint as
successor project. The installation of the tool is not necessary any
more, since we vendor the dependency directly into the repository. This
allows more reproducible lint results. All linters which currently pass
have been enabled. The gofmt and govet lint scripts have been removed
because they are now included in golangci-lint as well.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>