Add first crictl e2e test framework and suite
This adds a new Makefile target `test-e2e`, which uses the existing
ginkgo framework to run end-to-end tests for `crictl`.

The test framework and suite setups CRI-O in a tmp sandbox, whereas
every test can specify if a runtime is needed or not. This way it is
possible to create a new isolated environment where `crictl` can operate
on. Please be aware that external dependencies like `runc` and the CNI
plugins have to be existend on the system to acually work. The tests
have to be executed as root for now.

An example test suite for the `help` as well as the `info` command has
been added as well.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>