Update the default cri server
Currently, the default CRI server is dockershim. This
is a deprecated server with cri-tools. Also, does it make
sense to even have a default as the user should explicitly
define it.

This PR adds `containerd` and `cri-o` to the default list
and it checks each in order till one works. It also
adds some looging that the defaults are being used. This
at least provdes better usability to the user if they don't
configure the server.

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>