Update set configuration command
The config command currently implicitily sets a config option
when called with `option value`. The get is called with `--get` flag.
Also, the command doesn't allow more than one option to be set.
This commit updates it to use `--set` flag and be able to set
more than one options as `opt1=val1,opt2=val2`.

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>