Fix server connection timeout
The default server connection timeout is stated in documentation as
10s but in code it is 2s. Doc now updated.

It allows a timeout value of negative which is probably not a very good
setting as the system can assign it an unknown value which make be a
long/indefinite timeout. This now uses the default value when set to
negative..

Finally, there seemed to some bugs the way the timeout value is
retrieved and handled from config file and CLI. It doesn't seem to
be converted to seconds as expected.

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>