CI: bring back port forward test
This reverts commit 2688ed34bf00a7bc2fe718c748b0fbaf9ab6cb79,
and adds the following change on top:

-          sudo iptables -t nat -I POSTROUTING -s 127.0.0.1 ! -d 127.0.0.1 -j MASQUERADE
+          sudo iptables -t nat -I POSTROUTING -s 127.0.0.0/8 ! -d 127.0.0.0/8 -j MASQUERADE

The original DNS problem was caused by the fact that
systemd-resolved is listening at the IP address 127.0.0.53
and thus it was affected by the above rule.

Change the address to the address block (yes, the whole /8 block
is reserved for localhost addresses!) to fix this.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>