.github/workflows: fix GOROOT under sudo
It was found that using sudo (as in e.g. sudo make install) results in
wrong GOROOT setting, which leads to various weird compilation bugs.

Here is how GOROOT is changed:

-GOROOT="/opt/hostedtoolcache/go/1.15.4/x64"
+GOROOT="/opt/hostedtoolcache/go/1.14.11/x64"

Here are some of the errors caused by it:

	# runtime/internal/sys
	compile: version "go1.14.11" does not match go tool version "go1.15.4"
	# sync/atomic
	flag provided but not defined: -p

Let's use `sudo -E PATH=$PATH` for all cases which might call go.
This will make sure that proper go version is used (from $PATH), and
also that all the environment variables that it might need are correctly
set.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>