Makefile: avoid rebuilding binaries
It is a common scenario in which someone runs

	make
	sudo make install

Problem is, with the current Makefile, which has install-crictl depend
on a phony target crictl (and same for critest), it leads to making
crictl (and critest) targets twice. It is OK per se, since go is very
effective in caching dependencies and build artefacts, but it can be
problematic with sudo in case go is installed in some non-standard
path and/or requires some environment variables that sudo skips
(unless -E is used). As a result, binaries can be rebuilt again with
the wrong version of go, or some wrong environment, or maybe go binary
can't be found at all due to changed $PATH.

Anyway, to fix all this, let's create real (rather than phony) targets,
and have install depend on it.

At the same time, it would be good to have make always try to rebuild
the binaries, since it's not aware of their dependencies (but go does).
So, have the phony targets `crictl` and `critest` which run `make -B`
to force rebuild.

In other words, `make` (or `make binaries`, or `make crictl`) will now
always rebuild the binaries (same as before), but `make install` will
only try to build binaries if they are absent.

While at it:
 - put the built binaries into BUILD_BIN_DIR
 - consolidate and simplify some targets
 - remove non-existent targets from .PHONY
 - add some missing targets to .PHONY

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>