Modify hack/release.sh script to include sha256 sum files
We now add a *.sha256 file side by side to the tarball, which can be
uploaded as GitHub artifact. Beside this, the script now directly
outputs the Markdown table to lazy copy-paste. The script also now
includes the full file which has been build and not only it's name to
provide a more useful and verbose output.

Signed-off-by: Sascha Grunert <mail@saschagrunert.de>