Simplify test image build process for user images
This is the first set of a larger refactoring of our test images. We now
rely on a single script to build the test images for all architectures.
The script uses docker buildkit (buildx) to be independent from the
manifest tool.

We also build the images and manifests in the correct order, so that
image creation races are not possible any more.

Further on, I will migrate the other test images to the script as well.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>