Fix Containerd main branch CI for Windows
Due to the containerd makefile using $(PWD)/bin to determine where to place
the windows shim, it doesn't seem to like being run with powershell. I'm
not sure if it's just not inherited or what the reason is, but instead of
having the current dir replaced it simply would be an empty string. The
result of this is that the windows shim would get placed in /bin on the
machine instead of the bin in the current directory. 'make install' right
afterwards would then fail to find the file. To remedy this, use bash.exe
as the shell to run the CI. This is harmless and is intended to get the CI
working again for windows, but in Containerd using CURDIR instead of PWD may
be the solution to this problem.

This change additionally adds a small bit of logic to only build the windows
shim ourselves if we're running the CI for containerd < 1.6. In 1.6 (there's
only beta versions out so far but) the makefile has support for building
the shim itself. A simple 'make' or 'make binaries' invocation will build
the shim and place it in the same location as containerd and ctr.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>