Cleanup owners
We now move all `sig-node-approvers` to a synched alias in
`OWNERS_ALIASES` to simplify ownership. We also move heartlock,
tallclair, xlgao-zju and runcom to emeritus since they have not been
active in the past year.

Please mention on this PR if you still would like to be approver for
this project. :)

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>