userns: Skip tests if the host doesn't support idmap mounts
critest is used in projects like containerd, that test against older
distros (like AlmaLinux 8). In those distros, CI will fail when we
upgrade to runc 1.2.0.

With runc 1.1 those test don't fail because runc doesn't support idmap
mounts and the tests are skipped in that case. But with runc 1.2.0-rc.2,
that supports idmap mounts, the tests are not skipped but fail on
distros with older kernels that don't support idmap mounts.

This commit just tries to detect if the path used for the container
rootfs supports idmap mounts. To do that it uses the Status() message
from CRI with verbose param set to true. It parses the output that
containerd sets (it's quite unspecified that field), and otherwise
fallbacks to "/var/lib" as the path to test idmap mounts support.

Signed-off-by: Rodrigo Campos <rodrigoca@microsoft.com>