Disable runc integration tests due to AppArmor issue
All runc integration tests are failing in CI due to an AppArmor issue
in nested environments. The error occurs when runc tries to access
/proc/sys/net/ipv4/ip_unprivileged_port_start, which AppArmor
incorrectly interprets as trying to access /sys/... in detached mounts.

This is a known issue in runc that affects various container platforms.
The maintainers advise against downgrading as the current version fixes
multiple container escape vulnerabilities.

TODO: Re-enable these tests when a mitigation is found.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>